+++ Project/views/bulk-weather-import.ejs
@@ -127,34 +127,7 @@
 
     <% if (message) { %>
     <div class="message <%= message.type %>"><%= message.text %></div>
-    <% } %>
-
-    <form method="POST" action="/bulk-weather-import">
-      <div class="form-group">
-        <label for="location">Location:</label>
-        <input
-          type="text"
-          id="location"
-          name="location"
-          required
-          value="<%= location || 'Portland,OR' %>"
-          placeholder="City,State or Latitude,Longitude"
-        />
-      </div>
-      <div class="form-group">
-        <label for="date">Select Date:</label>
-        <input
-          type="date"
-          id="date"
-          name="date"
-          required
-          value="<%= date || '' %>"
-        />
-      </div>
-      <button type="submit">Get Weather Data</button>
-    </form>
-
-    <% if (weatherData && weatherData.days && weatherData.days[0]) { %>
+    <% } %> <% if (weatherData && weatherData.days && weatherData.days[0]) { %>
     <div class="import-summary">
       <h3>Weather Data Summary</h3>
       <p>Location: <%= weatherData.address %></p>
