+++ Project/views/import-csv.html
@@ -94,58 +94,8 @@
                     />
                     <label class="form-check-label" for="col-DATE">DATE</label>
                   </div>
-                  <div class="form-check">
-                    <input
-                      class="form-check-input"
-                      type="checkbox"
-                      value="LATITUDE"
-                      id="col-LATITUDE"
-                      checked
-                      disabled
-                    />
-                    <label class="form-check-label" for="col-LATITUDE"
-                      >LATITUDE</label
-                    >
-                  </div>
-                  <div class="form-check">
-                    <input
-                      class="form-check-input"
-                      type="checkbox"
-                      value="LONGITUDE"
-                      id="col-LONGITUDE"
-                      checked
-                      disabled
-                    />
-                    <label class="form-check-label" for="col-LONGITUDE"
-                      >LONGITUDE</label
-                    >
-                  </div>
                 </div>
                 <div class="col-md-4">
-                  <div class="form-check">
-                    <input
-                      class="form-check-input"
-                      type="checkbox"
-                      value="ELEVATION"
-                      id="col-ELEVATION"
-                      checked
-                      disabled
-                    />
-                    <label class="form-check-label" for="col-ELEVATION"
-                      >ELEVATION</label
-                    >
-                  </div>
-                  <div class="form-check">
-                    <input
-                      class="form-check-input"
-                      type="checkbox"
-                      value="NAME"
-                      id="col-NAME"
-                      checked
-                      disabled
-                    />
-                    <label class="form-check-label" for="col-NAME">NAME</label>
-                  </div>
                   <div class="form-check">
                     <input
                       class="form-check-input"
